+++ packages/admin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dyrected/admin
 
+## 2.5.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @dyrected/core@2.5.44
+  - @dyrected/sdk@2.5.44
+
 ## 2.5.43
 
 ### Patch Changes