+++ packages/db-mysql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dyrected/db-mysql
 
+## 2.5.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @dyrected/core@2.5.44
+
 ## 2.5.43
 
 ### Patch Changes